+++ next.config.ts
@@ -1,4 +1,3 @@
-import createMDX from '@next/mdx';
 import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
@@ -6,16 +5,4 @@ const nextConfig: NextConfig = {
   pageExtensions: ['js', 'jsx', 'md', 'mdx', 'ts', 'tsx'],
 };
 
-const withMDX = createMDX({
-  extension: /\.mdx?$/,
-  options: {
-    rehypePlugins: [['rehype-highlight']],
-    remarkPlugins: [
-      ['remark-frontmatter'],
-      ['remark-mdx-frontmatter'],
-      ['remark-gfm'],
-    ],
-  },
-});
-
-export default withMDX(nextConfig);
+export default nextConfig;