+++ src/features/post/ui/post-simple-card.tsx
@@ -1,4 +1,6 @@
-import { Frontmatter } from '@/entities/mdx/model/mdx.schema';
+'use client';
+
+import { Frontmatter } from '@/entities/markdown/model/markdown.schema';
 import { Button } from '@base-ui-components/react';
 import { format } from 'date-fns';
 import { Link2 } from 'lucide-react';
@@ -20,7 +22,7 @@ export default function PostSimpleCard({
       </div>
       <div className="col-span-9 flex flex-col gap-0.5">{summary}</div>
       <div className="col-span-3 flex-1 place-self-end">
-        <Link href={['/posts', ...path].join('/')}>
+        <Link href={['posts', ...path].join('/')}>
           <Button className="flex items-center place-self-end gap-2 px-4 py-2 text-sm text-white rounded bg-zinc-600 hover:bg-zinc-700 dark:bg-zinc-700 dark:hover:bg-zinc-600">
             <Link2 size={16} />
             Read More