+++ src/widgets/header.tsx
@@ -1,13 +1,17 @@
 'use client';
-import LocaleToggle from '@/shared/components/toggle/locale-toggle';
-import ThemeToggle from '@/shared/components/toggle/theme-toggle';
-import { useDetectScrolled } from '@/shared/hooks/useDetectScrolled';
-import clsx from 'clsx';
-import { Book, Mail, User } from 'lucide-react';
 import Link from 'next/link';
+
+import LocaleToggle from '@/shared/ui/toggle/locale-toggle';
+import ThemeToggle from '@/shared/ui/toggle/theme-toggle';
+import { useDetectScrolled } from '@/shared/hooks/useDetectScrolled';
+import { useLocaleStore } from '@/shared/stores/locale-store';
+
+import { Book, Mail, User } from 'lucide-react';
+import clsx from 'clsx';
 
 export default function Header() {
   const scrolled = useDetectScrolled();
+  const locale = useLocaleStore((state) => state.locale);
 
   const navButtonClasses = clsx(
     'flex items-center',
@@ -38,21 +42,29 @@ export default function Header() {
     <header className="flex items-center justify-center max-w-4xl h-16 sm:h-20 mx-auto">
       <div className={dynamicHeaderClasses}>
         <div className="ms-2">
-          <Link href="/" aria-label="Home" className={titleClasses}>
+          <Link href={`/${locale}`} aria-label="Home" className={titleClasses}>
             Chanho.dev
           </Link>
         </div>
         <div className="flex space-x-1">
-          <Link href="/about" aria-label="About" className={navButtonClasses}>
+          <Link
+            href={`/${locale}/about`}
+            aria-label="About"
+            className={navButtonClasses}
+          >
             <User size={16} />
             <span className="hidden md:inline">About</span>
           </Link>
-          <Link href="/posts" aria-label="Posts" className={navButtonClasses}>
+          <Link
+            href={`/${locale}/posts`}
+            aria-label="Posts"
+            className={navButtonClasses}
+          >
             <Book size={16} />
             <span className="hidden md:inline">Posts</span>
           </Link>
           <Link
-            href="/contact"
+            href={`/${locale}/contact`}
             aria-label="Contact"
             className={navButtonClasses}
           >